fix tests not running in circleci
